Remove unused utils and push more code into the Rust backend
