+++ lua/dash/health.lua
@@ -5,6 +5,14 @@ local function build_cli_path(dash_app_path)
   return (dash_app_path:gsub('(.)%/$', '%1')) .. require('libdash_nvim').DASH_APP_CLI_PATH
 end
 
+local function get_keys(tbl)
+  local keyset = {}
+  for k, _ in pairs(tbl) do
+    keyset[#keyset + 1] = k
+  end
+  return keyset
+end
+
 function M.check()
   -- health_start begins a new section
   local health_start = vim.fn['health#report_start']
@@ -33,7 +41,7 @@ function M.check()
   require('telescope').load_extension('dash')
 
   health_start('Configuration')
-  local config = require('dash.utils.config').config
+  local config = require('dash.config').config
   local dash_app_path = config.dash_app_path
   local cli_path = build_cli_path(dash_app_path)
   -- check if Dash.app CLI is executable at configured path
@@ -43,14 +51,14 @@ function M.check()
     health_error('Dash.app not found at configured path: ' .. dash_app_path)
 
     -- if not executable at configured path, check if executable at default path
-    local default_path = require('dash.utils.config').default_config.dash_app_path
+    local default_path = require('dash.config').default_config.dash_app_path
     if vim.fn.executable(build_cli_path(default_path)) == 0 then
       health_info('Dash.app found and executable at default path: ' .. default_path)
     end
   end
 
   -- check status of filtering by filetype
-  if not config.file_type_keywords or #(require('dash.utils.tables').get_keys(config.file_type_keywords)) == 0 then
+  if not config.file_type_keywords or #(get_keys(config.file_type_keywords)) == 0 then
     health_info(
       'Filetype filtering is disabled.\nconfig.file_type_keywords = ' .. vim.inspect(config.file_type_keywords)
     )
