+++ spec/dash/rust/backend_spec.lua
@@ -5,5 +5,7 @@ describe("require('libdash_nvim')", function()
     assert.is_true(libdash ~= nil)
     assert.is_true(libdash.query ~= nil)
     assert.are.equal('function', type(libdash.query))
+    assert.is_true(libdash.open_item ~= nil)
+    assert.are.equal('function', type(libdash.open_item))
   end)
 end)
